pyproject.toml
[project]
name = "silicon-sage"
version = "0.1.0"
description = "A Google ADK powered PC Builder Multi Agentic System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "pydantic>=2.12.4",
]
